Moved the text_util file to a subdirectory inside the task, and added its licencing statements.
