+++ parlai/tasks/natural_questions/utils/README.md
@@ -1,0 +1,2 @@
+A collection of extenral modules provieded along with the original NQ dataset.
+More information [here](https://github.com/google-research-datasets/natural-questions)
