+++ parlai/tasks/natural_questions/utils/text_utils.py
@@ -1,5 +1,18 @@
 #!/usr/bin/env python3
 
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+##########################################################################
+#
+#  Note:  this file is mostly copied from the utility functions provided
+#         with the the datatset itself. For more information check
+#         the original repository, linked blow:
+#   https://github.com/google-research-datasets/natural-questions
+#
+##########################################################################
+#
 # coding=utf-8
 # Copyright 2018 The Google AI Language Team Authors.
 #
